requirements.txt
numpy
pandas
scikit-learn
nltk
stanza
tweet-preprocessor
wordsegment
sentence-transformers
python-terrier
pulp